io: finished the set mode function. Has to tested yet. io: PGM read function fails.

+++ src/arch/avr/328/io.c
@@ -52,7 +52,7 @@ const unsigned short ROM BermudaPortToInput[] =
 
 const unsigned short ROM BermudaPortToMode[] =
 {
-        PIN_NOT_AVAILABLE,
+        (unsigned short)PIN_NOT_AVAILABLE,
         (unsigned short)BermudaGetAddressDDRB(),
         (unsigned short)BermudaGetAddressDDRC(),
         (unsigned short)BermudaGetAddressDDRD(),
